remove logging configuration
Logging configurations are highly application-specific and, thus, should
be set in the application's code [1]. This change improves the library
as follows:

* Both handlers of the library-specific logger are removed. Thus,
  messages are not written twice to the terminal anymore when the root
  logger has a terminal StreamHandler. Note that basicConfig is called
  if the root logger has no handlers and a module-level log function is
  called [2]. See [3] for details on message propagation.
* The levels of the root logger and the logger for asyncio are not
  changed anymore. A library should not do such things.
* Now by default, log messages are written to stderr instead of stdout.
  Also, by default only messages with level WARNING and up are written.
  Both are UNIX conventions.

The previously hard-coded log format string can be accessed with
KikClient.log_format() now.

KikClient's log_level argument is removed. Note that, in echo_bot.py,
the level was not explicitly set. So log_level's default value INFO was
in effect. In order to preserve its behaviour, that example script sets
the INFO level manually now.

[1] https://docs.python.org/3/howto/logging.html#configuring-logging-for-a-library
[2] https://docs.python.org/3/library/logging.html#logging.basicConfig
[3] https://docs.python.org/3/library/logging.html#logging.Logger.propagate