Refactor device_id and android_id
Since device_id and android_id are so tightly-bound to a username/
login session, and because there are potential problems surrounding
globally-configured values (i.e. account bans), I'd like to see the
API require the user to explicitly set these values.

I've removed the concept of a device_id or android_id "override", which
creates a more intuitive interface, while forcing the user to be
cognizant of these values.

While the user is able to manage these values however they like,
I've provided a suggested config-file-base approach in the example files.
Users can create a new configuration file using the
generate_credentials_file.py script, or make use of the credentials file
generated by register_client.py.

I realize that these API changes are not backwards-compatible. This
was a conscious decision, as it requires the user to reevaluate how
they are using device_id and android_id and make any necessary changes
at this time. The burden of code changes required should be fairly
minimal for any user of this library.